Fix limit handling.
Previously, we weren't actually using the locking the MVar provides,
which meant a race could cause too many jobs to run. Now we take the
lock, and only release it if:

a. We decide not to run now (too many recent messages sent).
b. We've added ourselves to the running list.

Which means that whenever someone takes the MVar, it reflects accurately
what is currently running.