Payload; re-exports; stable serialization; aliases
Payload: We now can carry an additional piece of data along with the
job. It will be passed to the `after` handler. Uses for this could
include any additional meta-data you need in order to store the results
of the message being sent (for example, if you store emails connected to
users, the user id could go in here).

Re-exports: We now re-export all that you need from `hworker`,
eliminating the confusing `create` hiding.

Stable serialization: This has two parts:
1. We don't use automatically generated ToJSON/FromJSON. This is really
important if we want to be able to change this in the future. Related to
that change, we now:
2. Store a 'v':1 in our serialized jobs. This is not used currently, but
if/when we change the format, it will allow us to migrate forward (ie,
we can dispatch on the 'v' and be able to read messages from the
previous format).

Aliases: As a minor convenience, add SESWorker and SESWorkerWith type aliases.