Payload is `Either Value a` for error tolerance
Previously, we needed to be able to deserialize the payload successfully
in order to have the job work. If the client of the worker changed what
their payload looked like, this would break all outstanding
jobs. Instead, we expose it as an optimistic deserialization that falls
back on the json structure if we can't make sense out of it. Ideally,
users will manage versioning well, but this means that that they can't
get burned by not realizing they should.