Catch exceptions raised in callbacks.
hworker-ses will guarantee that emails get sent out, and that the
callback gets called, but if the callback raises an exception, that's
not on hworker-ses to retry - it should just bail.

In particular, these callbacks could be buggy, and we don't want
exceptions thrown there to trigger retries, as we would end up spamming
people with emails.

The _only_ edge case here is that if there is an async exception that
kills the thread while the callback is being run, we'll catch it and
continue on. This obviously isn't great, and perhaps we should figure
out something else in the future, but this behavior is better than the
previous behavior.